Additional Story & UI
Added in some more states and some additional text to the story. Basic structure is all there now, it's just a matter of creating the content and finishing off the core loop.

Also removed the prison image and replaced it with some text & the littledumbco branding.

Clocking off for the night!