All Functional!
Well, that's all the states and relevant structure in place for what is a basic text adventure game.

The format was super simple to understand once you've done a couple - actually ended up finding writing the relevant story copy HARDER than the actual code.

Also managed to fix a couple of fixes that bothered me, one of those being the repetition of printing the state to the console. Took a few mins to work out, but found the print line in the end.

Pretty much did all of this part without the Unity course as I wanted to get my head wrapped around it and dug in.

Huzzah!